Move services_v2 into core/service_interface/
Pure move. Resolving which app provides a service is about to gain company —
a registry of services the router provides itself, and a client for calling
either kind — so it gets a folder rather than sitting loose in core/.

Filename unchanged so git records a rename; the only edited lines are imports
in five files.
